requirements.txt
Django==2.0.6
django-crispy-forms==1.7.2
django-mathfilters==0.4.0
pytz==2018.4
selenium==3.12.0
six==1.11.0
